botpy/requirements.txt
# Dependencies for Trust PVP Reinforcement Learning Bot
python-socketio>=5.8.0
numpy>=1.24.0
torch>=2.0.0
matplotlib>=3.7.0
tqdm>=4.65.0